+++ app/src/main/AndroidManifest.xml
@@ -11,11 +11,11 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Projecttrackingsystemandroidapp">
-        <activity android:name=".activity.EmployeeShowCommitsActivity"></activity>
-        <activity android:name=".activity.EmployeeInfoActivity" />
-        <activity android:name=".activity.AdminIndexActivity" />
-        <activity android:name=".activity.ManagerIndexActivity" />
-        <activity android:name=".activity.EmployeeIndexActivity" />
+        <activity android:name=".activity.employee.EmployeeShowCommitsActivity"></activity>
+        <activity android:name=".activity.employee.EmployeeInfoActivity" />
+        <activity android:name=".activity.admin.AdminIndexActivity" />
+        <activity android:name=".activity.manager.ManagerIndexActivity" />
+        <activity android:name=".activity.employee.EmployeeIndexActivity" />
         <activity android:name=".activity.LoginActivity" />
         <activity android:name=".activity.HomeActivity">
             <intent-filter>
