+++ app/src/main/java/com/selimhorri/pack/activity/employee/EmployeeIndexActivity.java
@@ -1,4 +1,4 @@
-package com.selimhorri.pack.activity;
+package com.selimhorri.pack.activity.employee;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -13,6 +13,7 @@ import android.view.MenuItem;
 import android.widget.Toast;
 
 import com.selimhorri.pack.R;
+import com.selimhorri.pack.activity.HomeActivity;
 import com.selimhorri.pack.pattern.adapter.EmployeeIndexCustomAdapter;
 import com.selimhorri.pack.service.EmployeeService;
 import com.selimhorri.pack.service.impl.dynmc.EmployeeServiceDynamicImpl;
