+++ app/src/main/java/com/selimhorri/pack/activity/employee/EmployeeInfoActivity.java
@@ -1,4 +1,4 @@
-package com.selimhorri.pack.activity;
+package com.selimhorri.pack.activity.employee;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -10,6 +10,7 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.selimhorri.pack.R;
+import com.selimhorri.pack.activity.HomeActivity;
 import com.selimhorri.pack.service.EmployeeService;
 import com.selimhorri.pack.service.impl.dynmc.EmployeeServiceDynamicImpl;
