+++ app/src/main/java/com/selimhorri/pack/activity/employee/EmployeeShowCommitsActivity.java
@@ -1,4 +1,4 @@
-package com.selimhorri.pack.activity;
+package com.selimhorri.pack.activity.employee;
 
 import androidx.appcompat.app.AppCompatActivity;
