+++ app/src/main/res/layout/activity_admin_index.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.AdminIndexActivity">
+    tools:context=".activity.admin.AdminIndexActivity">
 
     <Button
         android:id="@+id/button3"
