+++ app/src/main/res/layout/activity_employee_index.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.EmployeeIndexActivity">
+    tools:context=".activity.employee.EmployeeIndexActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerViewEmployeeProjectData"
