+++ app/src/main/res/layout/activity_employee_info.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.EmployeeInfoActivity">
+    tools:context=".activity.employee.EmployeeInfoActivity">
 
     <TextView
         android:id="@+id/textView5"
