+++ app/src/main/res/layout/activity_employee_show_commits.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.EmployeeShowCommitsActivity">
+    tools:context=".activity.employee.EmployeeShowCommitsActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
